+++ src/components/infopanel/components/ekspanderbartpanel/ekspanderbartpanel.less
